Правки относительно ТЗ Последовательность чисел относительно бесконечна: максимальное значение - это 9999.... Long.MAX_VALUE раз Элементы выделяются в шахматном порядке

+++ domain/src/main/java/me/kofesst/testtask/domain/NumberGenerator.kt
@@ -1,6 +1,8 @@
 package me.kofesst.testtask.domain
 
+import java.math.BigInteger
+
 interface NumberGenerator {
     val name: String
-    val numbersSequence: Sequence<Long>
+    val numbersSequence: Sequence<BigInteger>
 }
